Modernize. Use defines for Defaults keys: this is less error-prone. Also fixed an error.

+++ XeeControllerImageActions.m
@@ -11,7 +11,7 @@
 
 
 CGFloat XeeZoomLevels[]={0.03125,0.044,0.0625,0.09,0.125,0.18,0.25,0.35,0.5,0.70,1,1.5,2,3,4,6,8,11,16,23,32};
-int XeeNumberOfZoomLevels=21;
+NSInteger XeeNumberOfZoomLevels=21;
 
 
 @implementation XeeController (ImageActions)
